requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15.2
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-bootstrap4==22.2
django-environ==0.9.0
django-import-export==3.0.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pytz==2022.6
PyYAML==6.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.2
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
